+++ l10n_es_aeat/views/aeat_report_tax_mapping_view.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+    <data>
+
+        <record id="view_l10n_es_aeat_tax_line_tree" model="ir.ui.view">
+            <field name="name">l10n.es.aeat.tax.line.tree</field>
+            <field name="model">l10n.es.aeat.tax.line</field>
+            <field name="arch" type="xml">
+                <tree string="Tax lines" editable="bottom">
+                    <field name="field_number"/>
+                    <field name="name"/>
+                    <field name="amount"/>
+                    <button name="get_calculated_move_lines"
+                            type="object"
+                            help="Show journal items"
+                            icon="gtk-find"/>
+                </tree>
+            </field>
+        </record>
+
+    </data>
+</openerp>
